% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics.numeric}
\alias{calculate_statistics.numeric}
\title{Calculate Statistics for Numeric Data}
\usage{
\method{calculate_statistics}{numeric}(data)
}
\arguments{
\item{data}{A numeric vector for which the statistics will be calculated. Must contain at least 4 data points.}
}
\value{
A list with the following components:
\describe{
\item{min}{The minimum value of the data.}
\item{max}{The maximum value of the data.}
\item{mean}{The mean (average) of the data.}
\item{variance}{The variance (second data_moment) of the data.}
\item{sd}{The standard deviation of the data.}
\item{median}{The median of the data.}
\item{skewness}{The skewness of the data.}
\item{skewness_squared}{The square of the skewness.}
\item{kurtosis}{The kurtosis (excess kurtosis) of the data.}
}
}
\description{
This function calculates key statistics such as skewness, kurtosis, minimum, maximum, mean, variance, and standard deviation for a numeric dataset.
The skewness and kurtosis are computed using standard formulas, with kurtosis representing the Fisher kurtosis (excess kurtosis).
}
\examples{
data <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
calculate_statistics(data)
}
