% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cullenfrey_x}
\alias{cullenfrey_x}
\title{Run the Cullen-Frey Interactive Graph Application}
\usage{
cullenfrey_x(data)
}
\arguments{
\item{data}{A numeric vector, list, matrix  or dataframe that contains the datasets to be analyzed.}
}
\value{
Launches the Shiny web application.
}
\description{
This function launches a Shiny web application that allows users to interactively explore
the Cullen-Frey graph. It requires preprocessed data inputs for plotting, which should be
provided as arguments to the function.
}
