% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{data_moment}
\alias{data_moment}
\title{Calculate the k-th data_moment}
\usage{
data_moment(data, k)
}
\arguments{
\item{data}{A numeric vector representing the dataset.}

\item{k}{An integer representing the order of the data_moment to be calculated.}
}
\value{
A numeric value representing the k-th data_moment of the dataset.
}
\description{
This function calculates the k-th data_moment of a given dataset.
}
\examples{
data <- c(1, 2, 3, 4, 5)
data_moment(data, 2) # Calculate the second data_moment (The variance)
}
