% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{data_type.continuous}
\alias{data_type.continuous}
\title{Method for Continuous Data}
\usage{
\method{data_type}{continuous}(data)
}
\arguments{
\item{data}{A numeric vector containing the continuous dataset to process.}
}
\value{
A list of filtered data, theoretical points, and distribution lines.
}
\description{
Method for Continuous Data
}
