% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{data_type.discrete}
\alias{data_type.discrete}
\title{Method for Discrete Data}
\usage{
\method{data_type}{discrete}(data, ...)
}
\arguments{
\item{data}{A numeric vector containing the discrete dataset to process.}

\item{...}{Additional parameters.}
}
\value{
A list of filtered data, theoretical points, and distribution lines.
}
\description{
Method for Discrete Data
}
