% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{moment}
\alias{moment}
\title{Calculate the k-th Moment}
\usage{
moment(data, k)
}
\arguments{
\item{data}{A numeric vector representing the dataset.}

\item{k}{An integer representing the order of the moment to be calculated.}
}
\value{
A numeric value representing the k-th moment of the dataset.
}
\description{
This function calculates the k-th moment of a given dataset.
}
\examples{
data <- c(1, 2, 3, 4, 5)
moment(data, 2) # Calculate the second moment (The variance)
}
