% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate Statistics for a Given Dataset}
\usage{
calculate_statistics(data)
}
\arguments{
\item{data}{A numeric vector representing the dataset.}
}
\value{
A list containing calculated statistics such as minimum, maximum, median, skewness, and kurtosis.
}
\description{
This function is a generic method to calculate various statistics such as skewness, kurtosis, median, and range for a given numeric dataset.
}
\details{
The function dispatches based on the class of the input data and calculates statistics accordingly.
}
\examples{
data <- c(1, 2, 3, 4, 5)
calculate_statistics(data)
}
