% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{launch_cullen_frey_app}
\alias{launch_cullen_frey_app}
\title{Launch the Cullen-Frey Graph Shiny Application}
\usage{
launch_cullen_frey_app(data)
}
\arguments{
\item{data}{A numeric vector representing the dataset to be analyzed.
The dataset must contain at least 4 data points.}
}
\value{
This function does not return any value. It launches the Shiny application
displaying the Cullen-Frey graph.
}
\description{
This function calculates key statistics (skewness, kurtosis, etc.) for a given dataset and
prepares data for the Cullen-Frey graph. It then launches a Shiny application that
provides an interactive visualization of the Cullen-Frey graph based on the calculated statistics.
}
\examples{
\dontrun{
data <- rnorm(100)
launch_cullen_frey_app(data)
}
}
