% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_function.R
\name{method}
\alias{method}
\title{Dispatch Method for Statistical Calculation}
\usage{
method(data)
}
\arguments{
\item{data}{A numeric vector for which skewness and kurtosis will be calculated.}
}
\value{
The function returns a list containing the calculated skewness and kurtosis for the given data, depending on the method used.
}
\description{
This is a generic function for calculating statistical measures such as skewness and kurtosis.
It dispatches to specific methods like \code{method.sample} or \code{method.unbiased}, depending on the class of the input data.
}
\examples{
data <- rnorm(100)
result <- method(data)
}
