% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_function.R
\name{method.sample}
\alias{method.sample}
\title{Sample Method for Calculating Skewness and Kurtosis}
\usage{
\method{method}{sample}(data)
}
\arguments{
\item{data}{A numeric vector for which skewness and kurtosis will be calculated.}
}
\value{
A list containing skewness and kurtosis.
}
\description{
This function calculates skewness and kurtosis for a numeric dataset using the 'sample' method.
}
