% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_definition.R
\name{run_app}
\alias{run_app}
\title{Run the Cullen-Frey Interactive Graph Application}
\usage{
run_app(data)
}
\arguments{
\item{data}{A numeric vector representing the dataset to be analyzed.}
}
\value{
Launches the Shiny web application.
}
\description{
This function launches a Shiny web application that allows users to interactively explore
the Cullen-Frey graph. It requires preprocessed data inputs for plotting, which should be
provided as arguments to the function.
}
