% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_definition.R
\name{run_app}
\alias{run_app}
\title{Run the Cullen-Frey Interactive Graph Application}
\usage{
run_app(data, data_filtered, theoretical_points, lognormal_line, gamma_line)
}
\arguments{
\item{data}{A numeric vector representing the dataset to be analyzed.}

\item{data_filtered}{A data frame containing the filtered beta distribution data for plotting.}

\item{theoretical_points}{A data frame containing the theoretical distribution points.}

\item{lognormal_line}{A data frame containing the lognormal line data for plotting.}

\item{gamma_line}{A data frame containing the gamma line data for plotting.}
}
\value{
Launches the Shiny web application.
}
\description{
This function launches a Shiny web application that allows users to interactively explore
the Cullen-Frey graph. It requires preprocessed data inputs for plotting, which should be
provided as arguments to the function.
}
